requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
ampl-mp=3.1.0=h2cc385e_1006
asttokens=2.4.1=pyhd8ed1ab_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.8.30=hbcca054_0
cloudpickle=3.1.0=pyhd8ed1ab_1
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
cpython=3.13.0=py313hd8ed1ab_100
cuda-cudart=12.6.77=h5888daf_0
cuda-cudart_linux-64=12.6.77=h3f2d84a_0
cuda-nvrtc=12.6.77=hbd13f7d_0
cuda-nvtx=12.6.77=hbd13f7d_0
cuda-version=12.6=h7480c83_3
cudnn=9.3.0.75=h93bb076_0
debugpy=1.8.8=py313h46c70d0_0
decorator=5.1.1=pyhd8ed1ab_0
evdev=1.7.1=py313h536fd9c_3
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
filelock=3.16.1=pyhd8ed1ab_0
fsspec=2024.10.0=pyhff2d567_0
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py313h11186cd_2
icu=75.1=he02047a_0
importlib-metadata=8.5.0=pyha770c72_0
ipopt=3.14.16=h122424a_10
ipykernel=6.29.5=pyh3099207_0
ipyparallel=9.0.0=pyhff2d567_0
ipython=8.29.0=pyh707e725_0
jedi=0.19.2=pyhff2d567_0
jinja2=3.1.4=pyhd8ed1ab_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
keyutils=1.6.1=h166bdaf_0
krb5=1.21.3=h659f571_0
ld_impl_linux-64=2.43=h712a8e2_2
libabseil=20240722.0=cxx17_h5888daf_1
libblas=3.9.0=25_linux64_openblas
libboost=1.86.0=hb8260a3_2
libcblas=3.9.0=25_linux64_openblas
libcublas=12.6.3.3=hbd13f7d_1
libcufft=11.3.0.4=hbd13f7d_0
libcurand=10.3.7.77=hbd13f7d_0
libcusolver=11.7.1.2=hbd13f7d_0
libcusparse=12.5.4.2=hbd13f7d_0
libedit=3.1.20191231=he28a2e2_2
libexpat=2.6.4=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgomp=14.2.0=h77fa898_1
libhwloc=2.11.1=default_hecaa2ac_1000
libiconv=1.17=hd590300_2
liblapack=3.9.0=25_linux64_openblas
libmagma=2.8.0=h0af6554_0
libmagma_sparse=2.8.0=h0af6554_0
libmpdec=4.0.0=h4bc722e_0
libnvjitlink=12.6.77=hbd13f7d_1
libopenblas=0.3.28=pthreads_h94d23a6_1
libprotobuf=5.28.2=h5b01275_0
libscotch=7.0.4=h2fe6a88_5
libsodium=1.0.20=h4ab18f5_0
libspral=2024.05.08=h831f25b_3
libsqlite=3.47.0=hadc24fc_1
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtorch=2.5.1=cuda126_hd8d7199_303
libuuid=2.38.1=h0b41bf4_0
libuv=1.49.2=hb9d3cd8_0
libxml2=2.13.4=hb346dea_2
libzlib=1.3.1=hb9d3cd8_2
llvm-openmp=19.1.3=h024ca30_0
markupsafe=3.0.2=py313h8060acc_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
metis=5.1.0=hd0bcaf9_1007
mkl=2024.2.2=ha957f24_16
more-itertools=10.5.0=pyhd8ed1ab_0
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_0
mumps-include=5.7.3=ha770c72_5
mumps-seq=5.7.3=h27a6a8b_0
nccl=2.23.4.1=h52f6c39_2
ncurses=6.5=he02047a_1
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.4.2=pyhd8ed1ab_1
nlopt=2.8.0=py313h129aefb_2
numpy=2.1.3=py313h4bf6692_0
openssl=3.4.0=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pagmo=2.19.1=h4fefe70_3
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pip=24.3.1=pyh145f28c_0
platformdirs=4.3.6=pyhd8ed1ab_0
prompt-toolkit=3.0.48=pyha770c72_0
psutil=6.1.0=py313h536fd9c_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pybind11-abi=4=hd8ed1ab_3
pygments=2.18.0=pyhd8ed1ab_0
pygmo=2.19.7=py313h0d431c8_2
pynput=1.7.7=py313h78bf25f_1
python=3.13.0=h9ebbce0_100_cp313
python-dateutil=2.9.0=pyhd8ed1ab_0
python-xlib=0.33=pyhd8ed1ab_0
python_abi=3.13=5_cp313
pytorch=2.5.1=cuda126_py313he9a4f5b_303
pyyaml=6.0.2=py313h536fd9c_1
pyzmq=26.2.0=py313h8e95178_3
readline=8.2=h8228510_1
scipy=1.14.1=py313h27c5614_1
setuptools=75.3.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.7=h1b44611_2
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.3=pyh2585a3b_104
tbb=2021.13.0=h84d6215_0
tk=8.6.13=noxft_h4845f30_101
tornado=6.4.1=py313h536fd9c_1
tqdm=4.67.0=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
unixodbc=2.3.12=h661eb56_0
wcwidth=0.2.13=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.5=h3b0a872_6
zipp=3.21.0=pyhd8ed1ab_0
zlib=1.3.1=hb9d3cd8_2
zstd=1.5.6=ha6fb4c9_0
